server/requirements.txt
numpy==1.26.4
opencv-python
mediapipe==0.10.9
streamlit
imutils
scipy
scikit-learn
torch
torchvision
ultralytics
joblib
Pillow